+++ src/app.ts
@@ -2,6 +2,7 @@ import express from "express";
 import cron from "node-cron";
 import swaggerUi from "swagger-ui-express";
 import ExpressMongoSanitize from "express-mongo-sanitize";
+import cors from "cors";
 
 import { connection } from "./db/conn";
 import { getDataFromSpaceX } from "./utils/getDataFromSpaceXAPI";
@@ -9,7 +10,7 @@ import { router } from "./routes/routes";
 import { seed } from "./utils/seed";
 
 import { swaggerDef } from "./swaggerDef";
-import cors from "cors";
+import mongoose from "mongoose";
 
 const PORT = process.env.PORT || 3000;
 
@@ -25,25 +26,24 @@ app.use(ExpressMongoSanitize());
 // Adding middleware to accept json at body requests and limiting 10kb
 app.use(express.json({ limit: "10kb" }));
 
+app.use((_, res, next) => {
+  if (mongoose.connection.readyState !== 1)
+    return res.status(500).send("DB connection lost");
+  return next();
+});
+
 // Adding middleware to create routes to docs generated by swagger.
 app.use("/docs", swaggerUi.serve, swaggerUi.setup(swaggerDef));
 
 // Adding middleware to instance of all the routes.
 app.use("/", router);
 
-app.use(async (req, res, next) => {
-  try {
-    // Creating connection at mongoDb
-    await connection();
-    return next();
-  } catch (error) {
-    return res.status(500).send(error.message);
-  }
-});
-
 // Methods to start server.
 const start = async () => {
   try {
+    // create connection with mongoDB
+    await connection();
+
     // Feeding the db with spaceX public API
     await seed();
